Add new PeripheralDeviceConfigs
